requirements.txt
asgiref==3.9.1
contourpy==1.3.3
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.5
django-environ==0.12.0
fonttools==4.59.0
gunicorn==23.0.0
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
